Refactor – Rename settings.py → core.py across events, interfaces, mappers, repos (#896)
Rename-only change with no behavioral impact. The foundational base module
in each of the four affected packages is renamed from settings.py to core.py,
establishing the core.py convention uniformly across the framework layer.

- tiferet/events/settings.py → tiferet/events/core.py
- tiferet/interfaces/settings.py → tiferet/interfaces/core.py
- tiferet/mappers/settings.py → tiferet/mappers/core.py
- tiferet/repos/settings.py → tiferet/repos/core.py

All intra-package 'from .settings import' statements in submodules and
__init__.py files updated to 'from .core import'. Exported symbols unchanged.
