+++ tiferet/events/cli.py
@@ -6,7 +6,7 @@
 from typing import Optional, List
 
 # ** app
-from .settings import DomainEvent, a
+from .core import DomainEvent, a
 from ..domain import CliCommand
 from ..interfaces import CliService
 from ..mappers import CliCommandAggregate
