+++ tiferet/events/sqlite.py
@@ -7,7 +7,7 @@ from typing import List, Dict, Any, Sequence, Optional, Callable
 import sqlite3
 
 # ** app
-from .settings import DomainEvent, a
+from .core import DomainEvent, a
 from ..interfaces.sqlite import SqliteService
 
 # *** functions
