+++ tiferet/interfaces/middleware.py
@@ -7,7 +7,7 @@ from abc import abstractmethod
 from typing import Any, Callable, Dict
 
 # ** app
-from .settings import Service
+from .core import Service
 
 # *** interfaces
