+++ tiferet/mappers/__init__.py
@@ -29,7 +29,7 @@ __all__ = [
 ]
 
 # ** app
-from .settings import (
+from .core import (
     Aggregate,
     TransferObject,
 )
